requirements.txt
discord.py
python-dotenv
requests
GitPython
slackclient
Flask